Fixed MSVS2005 project file

+++ zkdecrypto/headers/zodiac.h
@@ -97,7 +97,7 @@ float IoC(const char *string)
 
 	if(!string) return 0;
 
-	length=strlen(string);
+	length=(int)strlen(string);
 	if(length<2) return 0;
 	memset(freqs,0,256*sizeof(int));
